! core: Remove usage of deprecated enclosingXXX API
From now on, widget type and layout type inference is abolished in order
not to rely on deprecated parent AST inspection in macros. This means that
instead of
w[Button] ~> { x => ... } // (v 1.0) or
w[Button] ~> (tweak ~ { x => ... }) // (v 2.0)
one should use
w[Button] ~> Tweak[Button] { x => ... }
A possible solution is to write a compiler plugin to perform the additional
inference.